fix: merge from multiplayer and mode support added
