+++ frontend/lib/models/character.dart
@@ -1,0 +1,4 @@
+class Character {
+  String name = "";
+  Character(this.name);
+}
